Feature: Add mongoose support and upgrade packages
Added mongoose database model generation option.

Upgraded dependent packages to the latest possible with support for
node 0.10x and then applied neceesary changes based on changes within
yeoman-generator.

Test generation is currently disabled because it does a "require"
on a generate file but yeoman-generator uses in-memory files so
require of a file not on disk does not work. This is being worked
on now.

Added more test cases to get index coverage up to ~96%.

Also includes Vagrantfile (for doing development on project) and
added docker images (see contrib) to be used in development.