Feature: Update x-handler support
Update the x-handler name to flow down to the name of the test files
as well. Because the structure assumes all methods for a given route
will be in the same file, it does not make sense to support handlers
at the individual path + verb level, as such that was removed.