Feature: Enable testing and yaml in generated tests
Update the generated package.json scripts test and cover to handle
invalid references test/ instead of tests/, to enable running the
generated test code. Also updates the lint reference to use the
correct structure instead of lib/**/*.js.

Update test templates to support yaml formatted API specs. Because the
enjoi library is for json schema, the yaml files need to be loaded and
parsed into objects for the tests to execute with a valid yaml formatted
API spec.