Chore: Resolve minor issues
githubUser is not required and generating incomplete entries within
package.json is not ideal. Updates the package.json to handle the
optional input and generate complete and valid file.

Provide optional argument for providing the app name.

Update resolution of the location of the api spec file relative
to the current directory or the new/existing project directory in
a consistent manner.

Separate the processing of the package.json file (if one exists)
from the setting of default values to simplify the method.